+++ packages/common/src/serializers/Profile.ts
@@ -1,9 +1,10 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
+
 import Profile from "@sentrei/common/models/Profile";
 
 const serializeProfile = (
   snap: firebase.firestore.DocumentSnapshot<Profile.Response>,
 ): Profile.Get => {
-  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
   const data = snap.data()!;
 
   return {